requirements.txt
speechrecognition
pyaudio
opencv-python
numpy>=1.24.0
pillow>=10.0.0
pandas
matplotlib
seaborn
ultralytics
torch
torchvision
pyserial
google-generativeai
pyyaml>=6.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
loguru>=0.7.0
streamlit>=1.18
streamlit-lottie
requests
plotly>=5.0.0